move the rest of the Write,WriteLine to partial class
